pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goblin"
version = "0.1.0"
description = "Automated job discovery bot with Slack integration and AWS Lambda deployment"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Nico"}]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "boto3>=1.34",
]

[project.urls]
Homepage = "https://github.com/dr-nico-f/goblin"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
goblin = "goblin.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
